Fix #66 Add FileType validation in parse API

+++ app.py
@@ -9,26 +9,36 @@ app = Flask(__name__)
 IS_PROD = os.environ.get("IS_PROD", False)
 
 
+def allowed_file(filename):
+    return '.' in filename and filename.rsplit('.', 1)[1].lower() == 'txt'
+
+
 @app.route('/parse-file', methods=['POST'])
 def parse_file():
     file = request.files['0']
-    filename = str(uuid.uuid4())
-    tmp_filepath = os.path.join("conversations", filename)
-    file.save(tmp_filepath)
-    try:
-        parsed_items, persons_list = get_parsed_file(tmp_filepath)
-        response = {
-            "success": True,
-            "chat": parsed_items,
-            "users": persons_list
-        }
-    except Exception as e:
+    if not allowed_file(file.filename):
         response = {
             "success": False,
-            "error_message": str(e)
+            "error_message": "Please upload a valid file!",
         }
+    else:
+        filename = str(uuid.uuid4())
+        tmp_filepath = os.path.join("conversations", filename)
+        file.save(tmp_filepath)
+        try:
+            parsed_items, persons_list = get_parsed_file(tmp_filepath)
+            response = {
+                "success": True,
+                "chat": parsed_items,
+                "users": persons_list
+            }
+        except Exception as e:
+            response = {
+                "success": False,
+                "error_message": str(e)
+            }
 
-    os.remove(tmp_filepath)
+        os.remove(tmp_filepath)
     return jsonify(response), 200
